feat: Add create table guard to cleanup location on Failure
